requirements.txt
openai
python-dotenv
instagrapi